etl_postgres_to_es/requirements.txt
backoff==1.10.0
certifi==2020.11.8
chardet==3.0.4
idna==2.10
psycopg2-binary==2.8.6
pydantic==1.7.2
python-decouple==3.3
requests==2.25.0
urllib3==1.26.1
